# docs/api/swagger.yaml
basePath: /api
definitions:
  article.CategoryCreateRequest:
    properties:
      name:
        maxLength: 50
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      store_id:
        type: integer
    required:
    - name
    - status
    type: object
  article.CategoryDeleteRequest:
    properties:
      id:
        type: integer
      store_id:
        type: integer
    required:
    - id
    type: object
  article.CategoryListRequest:
    properties:
      limit:
        maximum: 100
        minimum: 1
        type: integer
      name:
        type: string
      page:
        minimum: 1
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      store_id:
        type: integer
    required:
    - limit
    - page
    type: object
  article.CategoryUpdateRequest:
    properties:
      id:
        type: integer
      name:
        maxLength: 50
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      store_id:
        type: integer
    required:
    - id
    - name
    - status
    type: object
  article.CreateRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      image_id:
        type: integer
      show_type:
        enum:
        - 10
        - 20
        type: integer
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      store_id:
        type: integer
      title:
        maxLength: 300
        type: string
      virtual_views:
        type: integer
    required:
    - category_id
    - content
    - show_type
    - sort
    - status
    - title
    type: object
  article.DeleteRequest:
    properties:
      id:
        type: integer
      store_id:
        type: integer
    required:
    - id
    type: object
  article.ListRequest:
    properties:
      category_id:
        type: integer
      limit:
        maximum: 100
        minimum: 1
        type: integer
      page:
        minimum: 1
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      store_id:
        type: integer
      title:
        type: string
    required:
    - limit
    - page
    type: object
  article.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
  article.UpdateRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      image_id:
        type: integer
      show_type:
        enum:
        - 10
        - 20
        type: integer
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      store_id:
        type: integer
      title:
        maxLength: 300
        type: string
      virtual_views:
        type: integer
    required:
    - category_id
    - content
    - id
    - show_type
    - sort
    - status
    - title
    type: object
...
  response.Response:
    properties:
      cost:
        type: string
      data: {}
      errcode:
        type: integer
      message:
        type: string
      traceId:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  title: Zero API
  version: "1.0"
paths:
  /article/article/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.CreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建文章
      tags:
      - 文章管理
  /article/article/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.DeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除文章
      tags:
      - 文章管理
  /article/article/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.ListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/article.ListResult'
              type: object
      summary: 获取文章列表
      tags:
      - 文章管理
  /article/article/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.UpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新文章
      tags:
      - 文章管理
  /article/category/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.CategoryCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建文章分类
      tags:
      - 文章分类管理
  /article/category/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.CategoryDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除文章分类
      tags:
      - 文章分类管理
  /article/category/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.CategoryListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/article.ListResult'
              type: object
      summary: 获取文章分类列表
      tags:
      - 文章分类管理
  /article/category/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.CategoryUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新文章分类
      tags:
      - 文章分类管理
...
swagger: "2.0"